feat: add useHomeScope hook
